Correctif plugin custom

+++ post_processings.py
@@ -1,13 +1,15 @@
-from pelican import signals
+from pelican import signals, Pelican
+from typing import List
+from glob import glob
 import htmlmin
 import os
+import re
 
 
-def finalized(sender):
-    # Minification de sitemap.xml + suppression suffixes .html
+def minify_sitemap(sender: Pelican, filename) -> None:
     try:
-        with open(os.path.join(sender.output_path, 'sitemap.xml'), 'r+', encoding='utf-8') as f:
-            content = f.read().replace('.html', '')
+        with open(os.path.join(sender.settings.get('OUTPUT_PATH'), filename), 'r+', encoding='utf-8') as f:
+            content = f.read()
 
             f.seek(0)
             f.write(htmlmin.minify(content, remove_optional_attribute_quotes=False, remove_empty_space=True, remove_comments=True))
@@ -15,25 +17,48 @@ def finalized(sender):
     except FileNotFoundError:
         pass
 
-    # Suppression suffixes .html
-    filenames = [
-        sender.settings.get('ARCHIVES_SAVE_AS'),
-        sender.settings.get('CATEGORIES_SAVE_AS'),
-        sender.settings.get('TAGS_SAVE_AS'),
-    ]
+
+def make_url_seo_friendly(sender: Pelican, filenames: List[str]) -> None:
+    base_url = re.escape(sender.settings.get('SITEURL').removesuffix('/'))
+
+    index_urls_regex = re.compile(fr'({base_url}\S*/)index\.html')
+    page_urls_regex = re.compile(fr'({base_url}\S*)\.html')
 
     for filename in filenames:
         try:
-            with open(os.path.join(sender.output_path, filename), 'r+', encoding='utf-8') as f:
-                new_filename = filename.removesuffix('index.html').removesuffix('.html')
+            with open(os.path.join(sender.settings.get('OUTPUT_PATH'), filename), 'r+', encoding='utf-8') as f:
+                count = 0
 
-                content = f.read().replace(filename, new_filename)
+                (content, nsub) = index_urls_regex.subn(r'\1', f.read())
+
+                count += nsub
+
+                (content, nsub) = page_urls_regex.subn(r'\1', content)
+
+                count += nsub
+
+                if count == 0:
+                    continue
 
                 f.seek(0)
                 f.write(content)
                 f.truncate()
         except FileNotFoundError:
             pass
+
+
+def finalized(sender: Pelican) -> None:
+    filenames = []
+
+    for extension in ('rss', 'atom', 'xml', 'html'):
+        filenames.extend(glob(
+            f'**/*.{extension}',
+            root_dir=sender.settings.get('OUTPUT_PATH'),
+            recursive=True
+        ))
+
+    make_url_seo_friendly(sender, filenames)
+    minify_sitemap(sender, 'sitemap.xml')
 
 
 def register():
